# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Atzicay API",
        "description": "API documentation for Atzicay",
        "contact": {
            "email": "atzicay@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "api/atzicay/v1/"
        }
    ],
    "paths": {
...
        "/hangman": {
            "get": {
                "tags": [
                    "Hangman"
                ],
                "summary": "Get all hangman games",
                "description": "Retrieves all hangman games.",
                "operationId": "a204fb48fe71d19ad756d26bcd7980c9",
                "responses": {
                    "200": {
                        "description": "List of hangman games",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Hangman"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No hangman games found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Hangman"
                ],
                "summary": "Create a new hangman game",
                "description": "Creates a new hangman game.",
                "operationId": "54a2114110bdb64b1dada8ed7b31d772",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HangmanDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Hangman game created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Hangman"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/hangman/{id}": {
            "get": {
                "tags": [
                    "Hangman"
                ],
                "summary": "Get hangman game by ID",
                "description": "Retrieves a hangman game by its ID.",
                "operationId": "ca56fb948809b4c4a74330175c2fbfe8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the hangman game",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hangman game details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Hangman"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hangman game not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Hangman"
                ],
                "summary": "Update an existing hangman game",
                "description": "Updates an existing hangman game.",
                "operationId": "80e66d5bf6f3a95821654f2fae4b2ba2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the hangman game",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HangmanDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hangman game updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Hangman"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/memory-game/{id}": {
            "get": {
                "tags": [
                    "MemoryGame"
                ],
                "summary": "Get Memory Game by ID",
                "description": "Retrieves a memory game by its ID.",
                "operationId": "648eb6c3a85af9c547e60bc353813044",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the memory game",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Memory game found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemoryGame"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Memory game not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "MemoryGame"
                ],
                "summary": "Update Memory Game by ID",
                "description": "Updates a memory game by its ID.",
                "operationId": "bbe5a2a4182cce7b12eead24d023e4b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the memory game",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MemoryGameDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Memory game updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemoryGame"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/memory-game": {
            "post": {
                "tags": [
                    "MemoryGame"
                ],
                "summary": "Create a new Memory Game",
                "description": "Creates a new memory game.",
                "operationId": "3d7a6d17a492749c6dabf088cf1348fa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MemoryGameDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Memory game created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemoryGame"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HangmanDTO": {
                "title": "Hangman Data Transfer Object",
                "description": "DTO for Hangman game data",
                "required": [
                    "GameInstanceId",
                    "Word",
                    "Clue",
                    "Presentation"
                ],
                "properties": {
                    "GameInstanceId": {
                        "description": "The unique identifier for the game instance",
                        "type": "integer"
                    },
                    "Word": {
                        "description": "The word to guess in the hangman game",
                        "type": "string"
                    },
                    "Clue": {
                        "description": "A clue for the word to guess",
                        "type": "string"
                    },
                    "Presentation": {
                        "description": "The presentation of the word with guessed letters revealed",
                        "type": "string",
                        "enum": [
                            "A",
                            "F"
                        ]
                    }
                },
                "type": "object"
            },
            "MemoryGameDTO": {
                "title": "Memory Game DTO",
                "description": "Data transfer object for Memory Game",
                "properties": {
                    "GameInstanceId": {
                        "description": "The ID of the game instance",
                        "type": "integer"
                    },
                    "Mode": {
                        "description": "The mode of the game",
                        "type": "string",
                        "enum": [
                            "II",
                            "ID"
                        ]
                    },
                    "PathImg1": {
                        "description": "Path to the first image",
                        "type": "string"
                    },
                    "PathImg2": {
                        "description": "Path to the second image",
                        "type": "string",
                        "nullable": true
                    },
                    "DescriptionImg": {
                        "description": "Description of the image",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Hangman": {
                "title": "Hangman",
                "description": "Hangman entity schema",
                "properties": {
                    "GameInstanceId": {
                        "description": "The ID of the related game instance",
                        "type": "integer"
                    },
                    "Word": {
                        "description": "The word to guess in the hangman game",
                        "type": "string"
                    },
                    "Clue": {
                        "description": "The clue for the word",
                        "type": "string"
                    },
                    "Presentation": {
                        "description": "The presentation type of the hangman game",
                        "type": "string",
                        "enum": [
                            "A",
                            "F"
                        ]
                    }
                },
                "type": "object"
            },
            "MemoryGame": {
                "title": "MemoryGame",
                "description": "MemoryGame entity schema",
                "properties": {
                    "Id": {
                        "description": "Primary key of the MemoryGame",
                        "type": "integer"
                    },
                    "GameInstanceId": {
                        "description": "Foreign key referencing GameInstances",
                        "type": "integer"
                    },
                    "Mode": {
                        "description": "Mode of the memory game",
                        "type": "string"
                    },
                    "PathImg1": {
                        "description": "Path to the first image",
                        "type": "string"
                    },
                    "PathImg2": {
                        "description": "Path to the second image",
                        "type": "string",
                        "nullable": true
                    },
                    "DescriptionImg": {
                        "description": "Description of the images",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}